Add get_wikitext and make logging optional
Added a simple get_wikitext which can be used to fetch a page's wikitext.
This is really useful for Cirrus because it doesn't want to blindly create
pages.  It pretty frequently just verifies that pages have the text you
expect.

Making logging optional just allows you to clean up the terminal output.

Change-Id: Id0eb49fe3ac1aee3d320e9a1a6a52c7d89f7e2b9