Remove maximum gem versions for http libraries
The gems faraday and faraday_middleware have 1.x releases, and webmock is on version 3.x.

Removing the major version specifier allows this gem to be compatible with the latest versions of these industry-standard libraries. The test suite passes on the latest versions of all these gems.

I've also deployed this change to Wiki Education Dashboard, which is working smoothly.

Change-Id: I2cd703f4753a705c04c97fbde5c42734d84321ef